Restructured the repository
