+++ dwh/extract transform load/matlab/calcDuration.m
