+++ dwh/extract transform load/mongosh_aggregations.txt
